requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
urllib3<2
pandas
numpy
scikit-learn
xgboost
